requirements.txt
opencv-python>=4.8
numpy>=1.26.4,<2.0
Pillow>=10.3,<11
PyYAML>=6.0
pandas>=2.0,<2.2
openpyxl>=3.1
Jinja2>=3.1
cloudinary>=1.36
ebaysdk>=2.2.0
